# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Syncitt
# during dev, should point to your local machine
host: localhost:5095
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/getOffMarketList:
    # binds a127 app logic to a route
    x-swagger-router-controller: offMarket
    post:
      tags: ["Trader Module"]
      description: Get Category and Business List
      summary: Get Category and Business List
      # used as the method name of the controller
      operationId: getOffMarketList
      parameters:
        - name: body
          in: body
          description: page number and limit
          schema:
            $ref: "#/definitions/getOffMarketList"
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/sendMailForChat:
    # binds a127 app logic to a route
    x-swagger-router-controller: offMarket
    post:
      tags: ["Trader Module"]
      description: Get Category and Business List
      summary: Get Category and Business List
      # used as the method name of the controller
      operationId: sendMailForChat
      parameters:
        - name: body
          in: body
          description: page number and limit
          schema:
            $ref: "#/definitions/sendMailForChat"
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/updatePropertyClaimById:
    # binds a127 app logic to a route
    x-swagger-router-controller: offMarket
    post:
      tags: ["Property Module"]
      description: update Property Details By Id
      summary: update Property Details By Id
      # used as the method name of the controller
      operationId: updatePropertyClaimById
      parameters:
        - name: body
          in: body
          description: Get Property Details By Id
          schema:
            $ref: "#/definitions/updatePropertyClaimById"
          required: false
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  updatePropertyClaimById:
    type: object
    properties:
      id:
        type: string
      claimed:
        type: boolean
...
  getOffMarketList:
    type: object
    properties:
      limit:
        type: number
      locations:
        type: array
        items:
          type: object
      suburb_postcode:
        type: string
  sendMailForChat:
    type: object
    properties:
      limit:
        type: number
      locations:
        type: array
        items:
          type: object
      suburb_postcode:
        type: string
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string